+++ R/math.R
@@ -1,24 +1,11 @@
 # utility function for linear operations that can be done on coefs or
 #   evaluations directly.
 fun_math <- function(x, op, ...) {
-  attr_ret <- attributes(x)
-  dots <- list(...)
-  ret <- map(tf_evaluations(x), \(x) {
-    if (is.null(x)) {
-      return(NULL)
-    }
-    result <- do.call(op, c(list(x = x), dots))
-    if (allMissing(result)) NULL else result
-  })
+  ret <- map(tf_evaluations(x), \(evals) op_or_null(op, evals, ...))
   if (is_irreg(x)) {
-    ret <- map2(tf_arg(x), ret, \(x, y) {
-      if (is.null(y)) {
-        return(NULL)
-      }
-      list(arg = x, value = y)
-    })
+    ret <- irreg_pairs(tf_arg(x), ret)
   }
-  attributes(ret) <- attr_ret
+  attributes(ret) <- attributes(x)
   ret
 }
 #------------------------------------------------------------------------------